tests/wpt/web-platform-tests/preload/reflected-as-value.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var link = document.createElement("link");
  var values = {
    "Image": "image",
    "images": "",
    "scripT": "script",
    "style": "style",
    "": "",
    "foNt": "font",
    "foobar": "",
    "video": "video",
    "audio": "audio",
    "track": "track",
    "fetch": "fetch",
  };
  var keys = Object.keys(values);
  for (var i = 0; i < keys.length; ++i) {
    link.as = keys[i];
    assert_true(link.as == values[keys[i]]);
  }
}, "Make sure that the `as` value reflects only known values");
</script>
